# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 프로젝트 생성 공통기능 API
  version: 1.0.0
  description: 한국교원대학교 교수업적평가시스템 공통기능 1차 범위 HTTP API 계약
servers:
  - url: /
    description: 상대경로 /api 사용 계약
paths:
...
  /api/admin/users:
    get:
      operationId: searchUsers
      summary: 사용자 조건 조회
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 검증 오류
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
          description: 페이지 번호
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
          description: 페이지 크기
        - name: employeeNo
          in: query
          required: false
          schema:
            type: string
          description: employeeNo
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: name
        - name: organizationCodeFilter
          in: query
          required: false
          schema:
            type: string
          description: organizationCodeFilter
        - name: rankName
          in: query
          required: false
          schema:
            type: string
          description: rankName
        - name: employmentStatus
          in: query
          required: false
          schema:
            type: string
          description: employmentStatus
        - name: roleCodeFilter
          in: query
          required: false
          schema:
            type: string
          description: roleCodeFilter
        - name: systemUseYn
          in: query
          required: false
          schema:
            type: string
          description: systemUseYn
        - name: filter
          in: query
          required: false
          schema:
            type: string
      x-related-requirements:
        - REQ-010
        - REQ-011
        - REQ-002
        - REQ-015
        - REQ-079
        - REQ-068
        - REQ-070
        - REQ-076
        - REQ-077
      security:
        - SessionCookie: []
      x-roles:
        - R09
  /api/admin/users/{userId}/account:
    patch:
      operationId: updateUserAccount
      summary: 사용자 사용여부 변경
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 검증 오류
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: userId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserAccountRequest'
      x-related-requirements:
        - REQ-012
        - REQ-013
        - REQ-064
        - REQ-069
        - REQ-003
        - REQ-008
        - REQ-065
        - REQ-073
        - REQ-074
        - REQ-015
        - REQ-079
        - REQ-070
        - REQ-076
        - REQ-077
      security:
        - SessionCookie: []
      x-roles:
        - R09
      x-business-rules:
        - KORUS 원천정보 직접 수정 금지 요구의 입력 필드는 서버에서 검증한다
        - 인증된 R09 관리자만 사용자 사용여부 변경 요청을 수행할 수 있다
      x-side-effects:
        - 사용자 사용여부 변경 결과를 관련 table에 저장하고 변경 메타정보를 기록한다
      x-required-tests:
        - 'happy: 유효한 사용자 사용여부 변경 요청 -> 후속 조회에서 REQ-012 관련 table 변경값이 확인된다'
        - 'auth: 인증 없는 사용자 사용여부 변경 요청 -> 401 ApiError가 반환된다'
        - 'validation: 필수 field 누락 사용자 사용여부 변경 요청 -> 해당 field-level ApiError가 반환된다'
        - 'business: 사용자 사용여부 변경 업무 규칙 위반 -> 기존 상태가 변경되지 않는다'
        - 'side-effect: 사용자 사용여부 변경 성공 -> 변경 처리자와 처리일시가 관련 row에 남는다'
  /api/admin/users/{userId}/roles:
    get:
      operationId: listCurrentUserRoles
      summary: 사용자 현재 역할 조회
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 검증 오류
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: userId
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
          description: 페이지 번호
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
          description: 페이지 크기
      x-related-requirements:
        - REQ-033
        - REQ-034
        - REQ-015
        - REQ-079
        - REQ-070
        - REQ-076
        - REQ-077
        - REQ-082
      security:
        - SessionCookie: []
      x-roles:
        - R09
    patch:
      operationId: updateUserRoles
      summary: 사용자 업무 역할 변경
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 검증 오류
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: userId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRolesRequest'
      x-related-requirements:
        - REQ-014
        - REQ-030
        - REQ-031
        - REQ-035
        - REQ-008
        - REQ-065
        - REQ-073
        - REQ-015
        - REQ-079
        - REQ-070
        - REQ-076
        - REQ-077
        - REQ-082
      security:
        - SessionCookie: []
      x-roles:
        - R09
      x-business-rules:
        - 사용자 업무 역할 변경 요구의 입력 필드는 서버에서 검증한다
        - 인증된 R09 관리자만 사용자 업무 역할 변경 요청을 수행할 수 있다
      x-side-effects:
        - 사용자 업무 역할 변경 결과를 관련 table에 저장하고 변경 메타정보를 기록한다
      x-required-tests:
        - 'happy: 유효한 사용자 업무 역할 변경 요청 -> 후속 조회에서 REQ-014 관련 table 변경값이 확인된다'
        - 'auth: 인증 없는 사용자 업무 역할 변경 요청 -> 401 ApiError가 반환된다'
        - 'validation: 필수 field 누락 사용자 업무 역할 변경 요청 -> 해당 field-level ApiError가 반환된다'
        - 'business: 사용자 업무 역할 변경 업무 규칙 위반 -> 기존 상태가 변경되지 않는다'
        - 'side-effect: 사용자 업무 역할 변경 성공 -> 변경 처리자와 처리일시가 관련 row에 남는다'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: HttpOnly SameSite=Lax session cookie
  schemas:
    ApiResponse:
      type: object
      required:
        - success
        - meta
      properties:
        success:
          type: boolean
        data: {}
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          type: object
          additionalProperties: true
    ApiError:
      type: object
      required:
        - code
        - message
        - fields
      properties:
        code:
          type: string
        message:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
...
    UpdateUserAccountRequest:
      type: object
      required:
        - systemUseYn
        - changeReason
      properties:
        systemUseYn:
          $ref: '#/components/schemas/SystemUseYnEnum'
        changeReason:
          type: string
    UpdateUserRolesRequest:
      type: object
      required:
        - roleCodes
        - changeReason
      properties:
        roleCodes:
          type: array
          items:
            $ref: '#/components/schemas/RoleCodeEnum'
        validStartDate:
          type: string
          format: date
        validEndDate:
          type: string
          format: date
        changeReason:
          type: string
...
    RoleCodeEnum:
      type: string
      enum:
        - R01
        - R02
        - R03
        - R04
        - R05
        - R06
        - R07
        - R08
        - R09
    SystemUseYnEnum:
      type: string
      enum:
        - Y
        - N
...